5.14 book manage

+++ src/main/java/class4/spm/novelnook/pojo/FineInfo.java
@@ -27,7 +27,7 @@ public class FineInfo {
     @JsonFormat(pattern = "yyyy-MM-dd")
     Date returntime;
 
-    int fineamount;
+    double fineamount;
